Remove generator dependency on profile
