+++ generator/build.gradle
@@ -3,9 +3,9 @@
  */
 
 dependencies {
-    compile project(':profile')
     compile project(':common')
-    
+
+    compile 'org.jetbrains.kotlin:kotlin-stdlib:1.3.21'
     compile 'com.google.inject:guice:4.1.0'
     compile 'dk.brics.automaton:automaton:1.11-8'
     compile 'org.apache.commons:commons-csv:1.5'
