+++ generator/src/main/java/com/scottlogic/deg/generator/generation/GenerationConfigSource.java
@@ -1,14 +1,10 @@
 package com.scottlogic.deg.generator.generation;
 
 import com.scottlogic.deg.generator.config.detail.*;
-import com.scottlogic.deg.profile.v0_1.AtomicConstraintType;
 
 import java.nio.file.Path;
-import java.util.List;
 
 public interface GenerationConfigSource  {
-    List<AtomicConstraintType> getConstraintsToNotViolate();
-
     DataGenerationType getGenerationType();
     CombinationStrategyType getCombinationStrategyType();
     TreeWalkerType getWalkerType();
