+++ orchestrator/src/main/java/com/scottlogic/deg/orchestrator/guice/AllConfigSource.java
@@ -1,5 +1,6 @@
-package com.scottlogic.deg.generator.generation;
+package com.scottlogic.deg.orchestrator.guice;
 
+import com.scottlogic.deg.generator.generation.GenerationConfigSource;
 import com.scottlogic.deg.profile.guice.ProfileConfigSource;
 
 public interface AllConfigSource extends GenerationConfigSource, ProfileConfigSource {
