+++ orchestrator/src/main/java/com/scottlogic/deg/orchestrator/guice/AllModule.java
@@ -1,9 +1,7 @@
 package com.scottlogic.deg.orchestrator.guice;
 
 import com.google.inject.AbstractModule;
-import com.scottlogic.deg.generator.generation.AllConfigSource;
 import com.scottlogic.deg.generator.guice.GeneratorModule;
-import com.scottlogic.deg.orchestrator.validator.ConfigValidator;
 import com.scottlogic.deg.profile.guice.ProfileModule;
 
 public class AllModule extends AbstractModule {
