+++ orchestrator/src/main/java/com/scottlogic/deg/orchestrator/validator/ConfigValidator.java
@@ -2,7 +2,7 @@ package com.scottlogic.deg.orchestrator.validator;
 
 import com.google.inject.Inject;
 import com.scottlogic.deg.common.ValidationException;
-import com.scottlogic.deg.generator.generation.AllConfigSource;
+import com.scottlogic.deg.orchestrator.guice.AllConfigSource;
 import com.scottlogic.deg.generator.generation.GenerationConfigSource;
 import com.scottlogic.deg.generator.outputs.targets.OutputTargetValidationException;
 import com.scottlogic.deg.generator.utils.FileUtils;
