+++ orchestrator/src/main/java/com/scottlogic/deg/orchestrator/violate/ViolateCommandLine.java
@@ -5,7 +5,10 @@ import com.google.inject.Injector;
 import com.google.inject.Module;
 import com.scottlogic.deg.orchestrator.generate.GenerateCommandLine;
 import com.scottlogic.deg.orchestrator.guice.AllModule;
+import com.scottlogic.deg.profile.v0_1.AtomicConstraintType;
 import picocli.CommandLine;
+
+import java.util.List;
 
 @CommandLine.Command(
     name = "violate",
@@ -14,7 +17,8 @@ import picocli.CommandLine;
     parameterListHeading = "%nParameters:%n",
     optionListHeading = "%nOptions:%n",
     abbreviateSynopsis = true)
-public class ViolateCommandLine extends GenerateCommandLine {
+public class ViolateCommandLine extends GenerateCommandLine implements ViolateConfigSource {
+
     @Override
     public void run() {
         Module container = new ViolateModule(this);
@@ -24,4 +28,15 @@ public class ViolateCommandLine extends GenerateCommandLine {
 
         task.run();
     }
+
+    @CommandLine.Option(
+        names = {"--dont-violate"},
+        arity = "0..",
+        description = "Choose types of constraint should not be violated")
+    private List<AtomicConstraintType> constraintsToNotViolate;
+
+    public List<AtomicConstraintType> getConstraintsToNotViolate() {
+        return constraintsToNotViolate;
+    }
+
 }
